lab_lambda/requirements.txt
python-dotenv
sqlalchemy
PyMySQL
